refactor: restructuring the project
