+++ include/utils.h
@@ -1,0 +1,13 @@
+#ifndef UTILS
+#define UTILS
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+#define MAXSTRINGLENGTH 200
+
+void printErrorAndExit(char *errorMessage);
+
+
+#endif
