+++ src/common/utils.c
@@ -1,0 +1,9 @@
+#include "utils.h"
+
+// prints the message sent in the parameter and exit with error status
+void printErrorAndExit(char *errorMessage)
+{
+    puts(errorMessage);
+    exit(1);
+}
+
